deleting preview img along with original img in the deleteimg method

+++ backend/controller/local.class.php
@@ -158,7 +158,7 @@
 
 				$mod = new model_image();
 
-				if (DeletePic("img/" . $mod->GetImgLink($id_img))) {
+				if (DeletePic("img/" . $mod->GetImgLink($id_img)) && DeletePic("img/preview" . $mod->GetImgLink($id_img))) {
 					if ($mod->DeleteImg($id_img)) {
 						echo json_encode(['status' => 'success']);
 					}else{
